--- StarterDeckCode/StarterDecksData.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace JLPlugin.Data
-{
-    [System.Serializable]
-    public class StarterDecksData
-    {
-        public string name;
-        public List<string> cards;
-        public string iconTexture;
-    }
-}
