--- StarterDeckCode/StarterDecksDataMain.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace JLPlugin.Data
-{
-    [System.Serializable]
-    public partial class StarterDecksDataMain
-    {
-        public List<StarterDecksData> starterDecksData;
-    }
-}
